Move common utilities into `utilities/common`
