+++ src/__tests__/ApolloClient.ts
@@ -5,7 +5,7 @@ import { ApolloLink } from '../link/core/ApolloLink';
 import { HttpLink } from '../link/http/HttpLink';
 import { InMemoryCache } from '../cache/inmemory/inMemoryCache';
 import { makeReference } from '../cache/inmemory/helpers';
-import { stripSymbols } from '../utilities/util/stripSymbols';
+import { stripSymbols } from '../utilities/common/stripSymbols';
 import { withWarning } from '../util/wrap';
 import { ApolloClient } from '../';
 import { DefaultOptions } from '../ApolloClient';
