+++ src/__tests__/optimistic.ts
@@ -8,7 +8,7 @@ import { MutationQueryReducersMap } from '../core/types';
 import { Subscription } from '../util/Observable';
 import { ApolloClient } from '../';
 import { addTypenameToDocument } from '../utilities/graphql/transform';
-import { stripSymbols } from '../utilities/util/stripSymbols';
+import { stripSymbols } from '../utilities/common/stripSymbols';
 import { InMemoryCache } from '../cache/inmemory/inMemoryCache';
 import { makeReference } from '../cache/inmemory/helpers';
