+++ src/cache/inmemory/__tests__/writeToStore.ts
@@ -11,7 +11,7 @@ import gql from 'graphql-tag';
 
 import { storeKeyNameFromField } from '../../../utilities/graphql/storeUtils';
 import { addTypenameToDocument } from '../../../utilities/graphql/transform';
-import { cloneDeep } from '../../../utilities/util/cloneDeep';
+import { cloneDeep } from '../../../utilities/common/cloneDeep';
 import { StoreWriter } from '../writeToStore';
 import { defaultNormalizedCacheFactory } from '../entityCache';
 import { makeReference } from '../helpers';
