+++ src/core/__tests__/ObservableQuery.ts
@@ -16,7 +16,7 @@ import { ApolloClient } from '../../';
 
 import wrap from '../../util/wrap';
 import subscribeAndCount from '../../__tests__/utils/subscribeAndCount';
-import { stripSymbols } from '../../utilities/util/stripSymbols';
+import { stripSymbols } from '../../utilities/common/stripSymbols';
 import { ApolloError } from '../../errors/ApolloError';
 
 describe('ObservableQuery', () => {
