+++ src/core/__tests__/scheduler.ts
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag';
 
 import { InMemoryCache } from '../../cache/inmemory/inMemoryCache';
-import { stripSymbols } from '../../utilities/util/stripSymbols';
+import { stripSymbols } from '../../utilities/common/stripSymbols';
 
 import { QueryManager } from '../QueryManager';
 import { WatchQueryOptions } from '../../core/watchQueryOptions';
