+++ src/index.ts
@@ -106,12 +106,12 @@ export * from './utilities/graphql/fragments';
 export * from './utilities/graphql/getFromAST';
 export * from './utilities/graphql/transform';
 export * from './utilities/graphql/storeUtils';
-export * from './utilities/util/assign';
-export * from './utilities/util/canUse';
-export * from './utilities/util/cloneDeep';
-export * from './utilities/util/environment';
-export * from './utilities/util/errorHandling';
-export * from './utilities/util/isEqual';
-export * from './utilities/util/maybeDeepFreeze';
-export * from './utilities/util/mergeDeep';
-export * from './utilities/util/stripSymbols';
+export * from './utilities/common/assign';
+export * from './utilities/common/canUse';
+export * from './utilities/common/cloneDeep';
+export * from './utilities/common/environment';
+export * from './utilities/common/errorHandling';
+export * from './utilities/common/isEqual';
+export * from './utilities/common/maybeDeepFreeze';
+export * from './utilities/common/mergeDeep';
+export * from './utilities/common/stripSymbols';
