+++ src/utilities/common/__tests__/environment.ts
