+++ src/utilities/common/__tests__/stripSymbols.ts
