+++ src/utilities/common/assign.ts
