+++ src/utilities/common/maybeDeepFreeze.ts
