+++ src/utilities/common/stripSymbols.ts
